gimp-joke-980321
 * app/iscissors.c, app/tips_dialog.c, app/gradient.c,
        libgimp/gimp.c, plug-ins/AlienMap/AlienMap.c,
        plug-ins/bmp/bmp.c, plug-ins/fits/fitsrw.h,
        plug-ins/fits/fitsrw.c, plug-ins/flarefx/flarefx.c,
        plug-ins/gfig/gfig.c, plug-ins/gfli/gfli.c,
        plug-ins/gicon/gicon.c, plug-ins/gqbist/gqbist.c,
        plug-ins/gtm/gtm.c, plug-ins/hot/hot.c,
        plug-ins/ifscompose/ifscompose.c,
        plug-ins/ifscompose/ifscompose_utils.c
        plug-ins/max_rgb/max_rgb.c, plug-ins/nlfilt/nlfilt.c,
        plug-ins/pat/pat.c, plug-ins/pcx/pcx.c, plug-ins/rotate/rotate.c,
        plug-ins/script-fu/script-fu-server.c, plug-ins/snoise/snoise.c,
        plug-ins/threshold_alpha/threshold_alpha.c,
        plug-ins/zealouscrop/zealouscrop.c :gimp-joke-980321

        plug-ins/CML_explorer, plug-ins/autocrop,
        plug-ins/align_layers, plug-ins/blinds,
        plug-ins/bmp, plug-ins/megawidget: gimp-joke-980322

        Applied gimp-joke-980322-0.patch and gimp-joke-980321-0.patch
        from Yasuhiro SHIRASAKI <joke@awa.tohoku.ac.jp>. Mostly
        portability for DEC osf1's cc. Lots of MAIN();'s, trailing
        commas in enums, and guchar/gchar madness.

	Gimp now sends mail, cleans vinyl records,
	removes stubborn stains, julien fries, and more!

-adrian

+++ plug-ins/megawidget/megawidget.c
@@ -223,14 +223,14 @@ mw_radio_result(struct mwRadioGroup *rg) {
 
 GSList *
 mw_value_radio_group_new(GtkWidget *parent, gchar *name, 
-			 struct mwValueRadioGroup *rg, gint *var)
+			 struct mwValueRadioGroup *rg, glong *var)
 {
    GSList *lst = NULL;
    GtkWidget *frame;
    GtkWidget *vbox;
    GtkWidget *button;
    struct mwValueRadioGroup *c;
-   gint value;
+   glong value;
 
    if (name != NULL) {
      frame = gtk_frame_new(name);
@@ -707,12 +707,12 @@ ui_toggle_callback(GtkWidget *widget, gpointer data) {
 
 static void
 ui_value_toggle_callback(GtkWidget *widget, gpointer data) {
-   gint id;
+   glong id;
 
    /* Get radio button ID */
-   id=(gint)gtk_object_get_data(GTK_OBJECT(widget),"Radio_ID");
+   id=(glong)gtk_object_get_data(GTK_OBJECT(widget),"Radio_ID");
    if (GTK_TOGGLE_BUTTON (widget)->active) {
-     *(gint *)data= id;
+     *(glong *)data= id;
 #ifndef NO_PREVIEW
      if (do_preview && mw_do_preview!=NULL) (*mw_do_preview)(NULL);
 #endif
